a lot of

+++ customer/lib/data/models/fooddetail.dart
@@ -4,7 +4,7 @@ class FoodDetails {
   final Food food;
   final Map<String, int> selectedAddons;
   final int quantity;
-  final int finalPrice;
+  final double finalPrice;
 
   FoodDetails(this.food, this.selectedAddons, this.finalPrice, this.quantity);
